Moved experiment steps for Patching into separate file
