+++ varats-core/varats/provider/patch/patch_provider.py
@@ -32,78 +32,6 @@ from varats.utils.git_util import (
 )
 
 
-class ApplyPatch(actions.ProjectStep):
-    """Apply a patch to a project."""
-
-    NAME = "APPLY_PATCH"
-    DESCRIPTION = "Apply a Git patch to a project."
-
-    def __init__(self, project: VProject, patch: 'Patch') -> None:
-        super().__init__(project)
-        self.__patch = patch
-
-    # TODO: discuss signature
-    def __call__(self, _: tp.Any) -> StepResult:
-        try:
-            print(
-                f"Applying {self.__patch.shortname} to "
-                f"{self.project.source_of(self.project.primary_source)}"
-            )
-            apply_patch(
-                Path(self.project.source_of(self.project.primary_source)),
-                self.__patch.path
-            )
-
-        except ProcessExecutionError:
-            self.status = StepResult.ERROR
-
-        self.status = StepResult.OK
-
-        return StepResult.OK
-
-    def __str__(self, indent: int = 0) -> str:
-        return textwrap.indent(
-            f"* {self.project.name}: Apply patch "
-            f"{self.__patch.shortname}", " " * indent
-        )
-
-
-class RevertPatch(actions.ProjectStep):
-    """Revert a patch from a project."""
-
-    NAME = "REVERT_PATCH"
-    DESCRIPTION = "Revert a Git patch from a project."
-
-    def __init__(self, project, patch):
-        super().__init__(project)
-        self.__patch = patch
-
-    # TODO: discuss signature
-    def __call__(self, _: tp.Any) -> StepResult:
-        try:
-            print(
-                f"Reverting {self.__patch.shortname} on "
-                f"{self.project.source_of(self.project.primary_source)}"
-            )
-            revert_patch(
-                Path(self.project.source_of(self.project.primary_source)),
-                self.__patch.path
-            )
-
-        except ProcessExecutionError:
-            self.status = StepResult.ERROR
-
-        self.status = StepResult.OK
-
-        return StepResult.OK
-
-    def __str__(self, indent: int = 0) -> str:
-        return textwrap.indent(
-            f"* {self.project.name}: Revert patch "
-            f"{self.__patch.shortname}", " " * indent
-        )
-
-
 class Patch:
     """A class for storing a single project-specific Patch."""
